autograde winners; also general output format
the output format should be pretty simple and could definitely rely on the Award object class that can be found in find_winners.py, since it follows that structure. Might have to import this class to the other files to include things like presenter info on an award instance. i'm not sure of the best place to work with the human-readable and JSON outputs though